+++ gerris_erfolgs_tracker/integrations/google/drive_service.py
@@ -2,16 +2,13 @@ from __future__ import annotations
 
 from typing import Any
 
-from gerris_erfolgs_tracker.integrations.google.auth import (
-    BASE_SCOPES,
-    GOOGLE_SCOPE_DRIVE_READONLY,
-)
 from gerris_erfolgs_tracker.integrations.google.client import build_google_api_client
 from gerris_erfolgs_tracker.integrations.google.models import DriveFile, parse_google_datetime
+from gerris_erfolgs_tracker.integrations.google.scopes import BASE_SCOPES, GOOGLE_SCOPE_DRIVE
 
 DRIVE_API_BASE_URL = "https://www.googleapis.com/drive/v3"
 
-REQUIRED_SCOPES: tuple[str, ...] = (*BASE_SCOPES, GOOGLE_SCOPE_DRIVE_READONLY)
+REQUIRED_SCOPES: tuple[str, ...] = (*BASE_SCOPES, GOOGLE_SCOPE_DRIVE)
 
 
 def list_recent_files(